requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
pytz==2025.1
sqlparse==0.5.3